Refactor the products display div to be an independant component
